Maryia/Refactoring Notifications System (#4459)
* Maryia/NotificationStore draft [WIP]

* Maryia/prepared initial NotificationStore

* Maryia/started using NotificationStore

* Maryia/fix for successful launching after adding NotificationStore

* Maryia/changed custom_notifications & setClientNotifications interface

* Maryia/small fix for updated run-panel-store

* Maryia/connected NotificationStore to Bot & fixed handleClientNotifications

* Maryia/Put all reactions for notifications into NotificationStore + added should_show_popups

* Maryia/minor fix to pass circleci tests

* Maryia/removed extra setClientNotifications() calls

* Maryia/Showing popups in DTrader after it finishes mounting

* Maryia/started refactoring as per review comments
