+++ packages/account/src/Sections/Profile/PersonalDetails/personal-details.jsx
@@ -1224,7 +1224,7 @@ PersonalDetailsForm.propTypes = {
     history: PropTypes.object,
 };
 
-export default connect(({ client }) => ({
+export default connect(({ client, notifications }) => ({
     account_settings: client.account_settings,
     has_residence: client.has_residence,
     getChangeableFields: client.getChangeableFields,
@@ -1238,5 +1238,5 @@ export default connect(({ client }) => ({
     states_list: client.states_list,
     fetchResidenceList: client.fetchResidenceList,
     fetchStatesList: client.fetchStatesList,
-    refreshNotifications: client.refreshNotifications,
+    refreshNotifications: notifications.refreshNotifications,
 }))(withRouter(PersonalDetailsForm));
