+++ packages/core/src/App/Containers/RealAccountSignup/account-wizard.jsx
@@ -336,7 +336,7 @@ AccountWizard.propTypes = {
     residence_list: PropTypes.array,
 };
 
-export default connect(({ client, ui }) => ({
+export default connect(({ client, notifications, ui }) => ({
     account_settings: client.account_settings,
     is_fully_authenticated: client.is_fully_authenticated,
     realAccountSignup: client.realAccountSignup,
@@ -351,7 +351,7 @@ export default connect(({ client, ui }) => ({
     states_list: client.states_list,
     fetchStatesList: client.fetchStatesList,
     fetchResidenceList: client.fetchResidenceList,
-    refreshNotifications: client.refreshNotifications,
+    refreshNotifications: notifications.refreshNotifications,
     fetchFinancialAssessment: client.fetchFinancialAssessment,
     financial_assessment: client.financial_assessment,
 }))(AccountWizard);
