+++ packages/core/src/App/app.js
@@ -59,9 +59,10 @@ const initStore = notification_messages => {
     // TODO: Re-enable and update browser checking
     // OutdatedBrowser.init(root_store);!
     root_store.client.init();
-    root_store.ui.init(notification_messages);
     root_store.common.init();
     root_store.pushwoosh.init();
+    root_store.notifications.init();
+    root_store.ui.init(notification_messages);
 
     return root_store;
 };
