+++ packages/core/src/Stores/client-store.js
@@ -1,14 +1,11 @@
-import Cookies from 'js-cookie';
-import { action, computed, observable, reaction, runInAction, toJS, when } from 'mobx';
-import moment from 'moment';
 import {
-    redirectToLogin,
     getPropertyValue,
-    getUrlSmartTrader,
     getUrlBinaryBot,
+    getUrlSmartTrader,
     isDesktopOs,
     isEmptyObject,
     LocalStore,
+    redirectToLogin,
     setCurrencies,
     State,
     toMoment,
@@ -19,15 +16,17 @@ import {
     routes,
 } from '@deriv/shared';
 import { getLanguage, localize } from '@deriv/translations';
+import Cookies from 'js-cookie';
+import { action, computed, observable, reaction, runInAction, toJS, when } from 'mobx';
+import moment from 'moment';
 import { requestLogout, WS } from 'Services';
 import BinarySocketGeneral from 'Services/socket-general';
 import BinarySocket from '_common/base/socket_base';
 import * as SocketCache from '_common/base/socket_cache';
 import { isEuCountry, isMultipliersOnly, isOptionsBlocked } from '_common/utility';
 import BaseStore from './base-store';
-import { getClientAccountType, getAccountTitle } from './Helpers/client';
+import { getAccountTitle, getClientAccountType } from './Helpers/client';
 import { setDeviceDataCookie } from './Helpers/device';
-import { handleClientNotifications, clientNotifications } from './Helpers/client-notifications';
 import { buildCurrenciesList } from './Modules/Trading/Helpers/currency';
 
 const LANGUAGE_KEY = 'i18n_language';
@@ -66,10 +65,6 @@ export default class ClientStore extends BaseStore {
     @observable has_logged_out = false;
     @observable is_landing_company_loaded = false;
     @observable is_account_setting_loaded = false;
-
-    // TODO: Temporary variable. Remove after MX account closure has finished.
-    @observable client_notifications = clientNotifications;
-
     // this will store the landing_company API response, including
     // financial_company: {}
     // gaming_company: {}
@@ -605,31 +600,6 @@ export default class ClientStore extends BaseStore {
         return result;
     }
 
-    @computed
-    get custom_notifications() {
-        const notification_content = {
-            mx_mlt_notification: {
-                header: () => {
-                    if (this.has_malta_account || this.can_have_mlt_account) {
-                        return localize('Your Options account is scheduled to be closed');
-                    } else if (this.is_uk) {
-                        return localize('Your Gaming account is scheduled to be closed');
-                    }
-                    return localize('Your account is scheduled to be closed');
-                },
-                main: () => {
-                    if (this.has_malta_account || this.can_have_mlt_account) {
-                        return localize('Withdraw all funds from your Options account.');
-                    } else if (this.is_uk) {
-                        return localize('Please withdraw all your funds as soon as possible.');
-                    }
-                    return localize('Please proceed to withdraw your funds before 30 November 2021.');
-                },
-            },
-        };
-        return notification_content;
-    }
-
     // Manual list of MLT countries during MLT/MX account removal.
     // Also needed to check onboarding modal text for specific country.
     @computed
@@ -1095,10 +1065,10 @@ export default class ClientStore extends BaseStore {
             LocalStore.setObject(storage_key, JSON.parse(JSON.stringify(this.accounts)));
         }
         this.selectCurrency(currency);
-        this.root_store.ui.removeNotificationMessage({
+        this.root_store.notifications.removeNotificationMessage({
             key: 'currency',
         });
-        this.root_store.ui.removeNotificationByKey({
+        this.root_store.notifications.removeNotificationByKey({
             key: 'currency',
         });
         await this.init();
@@ -1167,16 +1137,6 @@ export default class ClientStore extends BaseStore {
         return filtered_list.length > 0 && filtered_list.every(acc => acc.is_disabled);
     };
 
-    getRiskAssessment = () => {
-        if (!this.account_status) return false;
-
-        const status = this.account_status?.status;
-
-        return this.isAccountOfType('financial')
-            ? /(financial_assessment|trading_experience)_not_complete/.test(status)
-            : /financial_assessment_not_complete/.test(status);
-    };
-
     shouldCompleteTax = () => {
         if (!this.isAccountOfType('financial')) return false;
 
@@ -1208,37 +1168,25 @@ export default class ClientStore extends BaseStore {
     async switchAccount(loginid) {
         this.setPreSwitchAccount(true);
         this.setIsLoggingIn(true);
-        this.root_store.ui.removeNotifications(true);
-        this.root_store.ui.removeAllNotificationMessages(true);
+        this.root_store.notifications.removeNotifications(true);
+        this.root_store.notifications.removeAllNotificationMessages(true);
         this.setSwitched(loginid);
         this.responsePayoutCurrencies(await WS.authorized.payoutCurrencies());
     }
 
     @action.bound
     async resetVirtualBalance() {
-        this.root_store.ui.removeNotificationByKey({ key: 'reset_virtual_balance' });
-        this.root_store.ui.removeNotificationMessage({ key: 'reset_virtual_balance', should_show_again: true });
+        this.root_store.notifications.removeNotificationByKey({ key: 'reset_virtual_balance' });
+        this.root_store.notifications.removeNotificationMessage({
+            key: 'reset_virtual_balance',
+            should_show_again: true,
+        });
         await WS.authorized.topupVirtual();
     }
 
     @action.bound
     switchEndSignal() {
         this.switch_broadcast = false;
-    }
-
-    @action.bound
-    refreshNotifications() {
-        this.root_store.ui.removeNotifications(true);
-        this.root_store.ui.removeAllNotificationMessages();
-        const client = this.accounts[this.loginid];
-        const { has_missing_required_field } = handleClientNotifications(
-            client,
-            this,
-            this.root_store.ui,
-            this.root_store.modules.cashier,
-            this.root_store.common
-        );
-        this.setHasMissingRequiredField(has_missing_required_field);
     }
 
     /**
@@ -1273,7 +1221,7 @@ export default class ClientStore extends BaseStore {
         this.setLoginId(LocalStore.get('active_loginid'));
         this.setAccounts(LocalStore.getObject(storage_key));
         this.setSwitched('');
-        let client = this.accounts[this.loginid];
+        const client = this.accounts[this.loginid];
         // If there is an authorize_response, it means it was the first login
         if (authorize_response) {
             // If this fails, it means the landing company check failed
@@ -1307,37 +1255,6 @@ export default class ClientStore extends BaseStore {
                 window.location.replace(urlForLanguage(authorize_response.authorize.preferred_language));
             }
         }
-
-        /**
-         * Set up reaction for account_settings, account_status, is_p2p_visible
-         */
-        reaction(
-            () => [
-                this.account_settings,
-                this.account_status,
-                this.landing_companies,
-                this.root_store.modules?.cashier?.general_store?.is_p2p_visible,
-                this.root_store.common?.selected_contract_type,
-                this.is_eu,
-            ],
-            () => {
-                client = this.accounts[this.loginid];
-                if (Object.keys(this.landing_companies).length > 0) {
-                    this.root_store.ui.removeNotifications();
-                    this.root_store.ui.removeAllNotificationMessages();
-                    const { has_missing_required_field } = handleClientNotifications(
-                        client,
-                        this,
-                        this.root_store.ui,
-                        this.root_store.modules.cashier,
-                        this.root_store.common
-                    );
-                    if (client && !client.is_virtual) {
-                        this.setHasMissingRequiredField(has_missing_required_field);
-                    }
-                }
-            }
-        );
 
         this.selectCurrency('');
 
@@ -1384,7 +1301,7 @@ export default class ClientStore extends BaseStore {
     responseWebsiteStatus(response) {
         this.website_status = response.website_status;
         if (this.website_status.message && this.website_status.message.length) {
-            this.root_store.ui.addNotificationMessage({
+            this.root_store.notifications.addNotificationMessage({
                 key: 'maintenance',
                 header: localize('Site is being updated'),
                 message: localize(this.website_status.message),
@@ -1392,7 +1309,7 @@ export default class ClientStore extends BaseStore {
                 is_persistent: true,
             });
         } else {
-            this.root_store.ui.removeNotificationMessage({
+            this.root_store.notifications.removeNotificationMessage({
                 key: 'maintenance',
             });
         }
@@ -1453,11 +1370,6 @@ export default class ClientStore extends BaseStore {
     @action.bound
     setSwitched(switched) {
         this.switched = switched;
-    }
-
-    @action.bound
-    setHasMissingRequiredField(has_missing_required_field) {
-        this.has_missing_required_field = has_missing_required_field;
     }
 
     /**
@@ -1535,7 +1447,7 @@ export default class ClientStore extends BaseStore {
 
     handleNotFoundLoginId() {
         // Logout if the switched_account doesn't belong to any loginid.
-        this.root_store.ui.addNotificationMessage({
+        this.root_store.notifications.addNotificationMessage({
             message: localize('Could not switch to default account.'),
             type: 'danger',
         });
@@ -1556,7 +1468,7 @@ export default class ClientStore extends BaseStore {
             }
 
             // Send a toast message to let the user know we can't switch his account.
-            this.root_store.ui.addNotificationMessage({
+            this.root_store.notifications.addNotificationMessage({
                 message: localize('Switching to default account.'),
                 type: 'info',
             });
@@ -1608,8 +1520,8 @@ export default class ClientStore extends BaseStore {
             () => {
                 // Remove real account notifications upon switching to virtual
                 if (this.accounts[this.switched]?.is_virtual) {
-                    this.root_store.ui.removeNotifications(true);
-                    this.root_store.ui.removeAllNotificationMessages();
+                    this.root_store.notifications.removeNotifications(true);
+                    this.root_store.notifications.removeAllNotificationMessages();
                 }
 
                 this.switchAccountHandler();
@@ -1618,39 +1530,11 @@ export default class ClientStore extends BaseStore {
     }
 
     @action.bound
-    resetVirtualBalanceNotification(loginid) {
-        if (!this.is_logged_in) return;
-        if (!this.accounts[loginid].is_virtual) return;
-        const min_reset_limit = 1000;
-        const max_reset_limit = 999000;
-        const balance = parseInt(this.accounts[loginid].balance);
-
-        // Display notification message to user with virtual account to reset their balance
-        // if the balance is less than equals to 1000 or more than equals to 999000
-        if (balance <= min_reset_limit || balance >= max_reset_limit) {
-            let message = localize(
-                'Your demo account balance is low. Reset your balance to continue trading from your demo account.'
-            );
-            if (balance >= max_reset_limit)
-                message = localize(
-                    'Your demo account balance has reached the maximum limit, and you will not be able to place new trades. Reset your balance to continue trading from your demo account.'
-                );
-            this.root_store.ui.addNotificationMessage(
-                clientNotifications({}, { resetVirtualBalance: this.resetVirtualBalance, message })
-                    .reset_virtual_balance
-            );
-        } else {
-            this.root_store.ui.removeNotificationByKey({ key: 'reset_virtual_balance' });
-            this.root_store.ui.removeNotificationMessage({ key: 'reset_virtual_balance', should_show_again: true });
-        }
-    }
-
-    @action.bound
     setBalanceActiveAccount(obj_balance) {
         if (this.accounts[obj_balance?.loginid] && obj_balance.loginid === this.loginid) {
             this.accounts[obj_balance.loginid].balance = obj_balance.balance;
             if (this.accounts[obj_balance.loginid].is_virtual) {
-                this.resetVirtualBalanceNotification(obj_balance.loginid);
+                this.root_store.notifications.resetVirtualBalanceNotification(obj_balance.loginid);
             }
             this.resetLocalStorageValues(this.loginid);
         }
@@ -1746,7 +1630,7 @@ export default class ClientStore extends BaseStore {
         runInAction(async () => {
             this.responsePayoutCurrencies(await WS.payoutCurrencies());
         });
-        this.root_store.ui.removeAllNotificationMessages(true);
+        this.root_store.notifications.removeAllNotificationMessages(true);
         this.syncWithLegacyPlatforms(this.loginid, this.accounts);
         this.cleanupRealityCheck();
     }
