+++ packages/trader/src/Stores/Modules/Portfolio/portfolio-store.js
@@ -240,7 +240,7 @@ export default class PortfolioStore extends BaseStore {
                         ...response.error,
                     });
                 } else {
-                    this.root_store.ui.addNotificationMessage(contractCancelled());
+                    this.root_store.notifications.addNotificationMessage(contractCancelled());
                 }
             });
         }
@@ -278,7 +278,7 @@ export default class PortfolioStore extends BaseStore {
                 sell_price: response.sell.sold_for,
                 transaction_id: response.sell.transaction_id,
             };
-            this.root_store.ui.addNotificationMessage(
+            this.root_store.notifications.addNotificationMessage(
                 contractSold(this.root_store.client.currency, response.sell.sold_for)
             );
         }
