+++ packages/trader/src/Stores/Modules/Trading/trade-store.js
@@ -1098,6 +1098,7 @@ export default class TradeStore extends BaseStore {
         if (this.is_trade_component_mounted && this.should_skip_prepost_lifecycle) {
             return;
         }
+        this.root_store.notifications.setShouldShowPopups(false);
         this.onPreSwitchAccount(this.preSwitchAccountListener);
         this.onSwitchAccount(this.accountSwitcherListener);
         this.onLogout(this.logoutListener);
@@ -1107,7 +1108,8 @@ export default class TradeStore extends BaseStore {
         this.setChartStatus(true);
         runInAction(async () => {
             this.is_trade_component_mounted = true;
-            this.prepareTradeStore();
+            await this.prepareTradeStore();
+            this.root_store.notifications.setShouldShowPopups(true);
         });
         // TODO: remove this function when the closure of MX and MLT accounts is completed.
         this.manageMxMltRemovalNotification();
@@ -1115,33 +1117,21 @@ export default class TradeStore extends BaseStore {
 
     @action.bound
     manageMxMltRemovalNotification() {
-        const client_notifications = this.root_store.client.client_notifications;
+        const { addNotificationMessage, client_notifications, notification_messages, unmarkNotificationMessage } =
+            this.root_store.notifications;
         const get_notification_messages = JSON.parse(localStorage.getItem('notification_messages'));
-        const is_logged_in = this.root_store.client.is_logged_in;
-        const has_iom_account = this.root_store.client.has_iom_account;
-        const has_malta_account = this.root_store.client.has_malta_account;
-        this.root_store.ui.unmarkNotificationMessage({ key: 'close_mx_mlt_account' });
+        const { has_iom_account, has_malta_account, is_logged_in } = this.root_store.client;
+        unmarkNotificationMessage({ key: 'close_mx_mlt_account' });
         if (get_notification_messages !== null && is_logged_in && (has_iom_account || has_malta_account)) {
             when(
-                () => this.root_store.client.is_logged_in && this.root_store.ui.notification_messages.length === 0,
+                () => is_logged_in && notification_messages.length === 0,
                 () => {
                     const hidden_close_account_notification =
                         parseInt(localStorage.getItem('hide_close_mx_mlt_account_notification')) === 1;
                     const should_retain_notification =
                         (has_iom_account || has_malta_account) && !hidden_close_account_notification;
                     if (should_retain_notification) {
-                        const mx_mlt_custom_header_reaction =
-                            this.root_store.client.custom_notifications.mx_mlt_notification.header();
-                        const mx_mlt_custom_content_reaction =
-                            this.root_store.client.custom_notifications.mx_mlt_notification.main();
-                        this.root_store.ui.addNotificationMessage(
-                            client_notifications(
-                                this.root_store.ui,
-                                {},
-                                mx_mlt_custom_header_reaction,
-                                mx_mlt_custom_content_reaction
-                            ).close_mx_mlt_account
-                        );
+                        addNotificationMessage(client_notifications.close_mx_mlt_account);
                     }
                 }
             );
@@ -1178,8 +1168,8 @@ export default class TradeStore extends BaseStore {
         this.root_store.modules.contract_trade.onUnmount();
         this.refresh();
         this.resetErrorServices();
-        if (this.root_store.ui.is_notifications_visible) {
-            this.root_store.ui.toggleNotificationsModal();
+        if (this.root_store.notifications.is_notifications_visible) {
+            this.root_store.notifications.toggleNotificationsModal();
         }
         if (this.prev_chart_layout) {
             this.prev_chart_layout.is_used = false;
