+++ packages/trader/src/Stores/index.js
@@ -9,5 +9,6 @@ export default class RootStore {
         this.gtm = core_store.gtm;
         this.rudderstack = core_store.rudderstack;
         this.pushwoosh = core_store.pushwoosh;
+        this.notifications = core_store.notifications;
     }
 }
